refactor: reorganize navigation utilities into separate modules
- Move navigation utilities from keyboardNavigation/helpers.ts to utils/navigation/
- Split utilities into three focused modules:
  - domHelpers.ts: DOM-related utilities (getElementId, getCommentKey)
  - lineHelpers.ts: Line type and content utilities
  - positionHelpers.ts: Position finding utilities from App.tsx
- Remove index.ts barrel export in favor of direct imports for better tree shaking
- Update all imports to use specific module paths
