+++ src/client/hooks/keyboardNavigation/index.ts
@@ -1,4 +1,3 @@
 export * from './types';
-export * from './helpers';
 export * from './navigationFilters';
 export * from './scrollUtils';
